+++ markmeld/const.py
@@ -1,0 +1,3 @@
+# Constants
+
+PKG_NAME = "markmeld"
