+++ markmeld/exceptions.py
@@ -3,3 +3,9 @@ class TargetError(Exception):
     There was a problem with a target
     """
     pass
+
+class ConfigError(Exception):
+	"""
+	There was a problem with the configuration file
+	"""
+	pass
