+++ tests/some_data.yaml
@@ -1,0 +1,2 @@
+number_of_things: 15
+name_of_person: The sundance kid
