+++ tests/some_text.md
@@ -1,0 +1,7 @@
+---
+text_property: text_property_value
+---
+
+# Text content
+
+Here is some text
