+++ tests/test_markmeld.py
@@ -15,9 +15,13 @@ def compare_to_file(file, string_to_compare):
         assert file_contents == string_to_compare
 
 
+def test_cli():
+    from markmeld.cli import main
+    main(test_args={"config":"tests/_markmeld.yaml"})
+
 def test_MarkdownMelder_demo():
     cfg = markmeld.load_config_file("demo/_markmeld.yaml")
-    cmd_data = markmeld.populate_cmd_data(cfg, "default", {})
+    # cmd_data = markmeld.populate_cmd_data(cfg, "default", {})
     x = markmeld.MarkdownMelder(cfg)
 
     res = x.build_target("default", print_only=True)
